only update tmux status when necessary
having a status-interval of 1 seemed to be keeping my cpu at a steady 3%
(perhaps just in sync with conky?). So intead, set a really large
interval (setting to 0 didn't seem to solve the cpu usage), and update
various mappings to refresh the status line manually when executed. This
change allows my cpu to drop down to 1% when idle now.