use xmobar for workspace portion of xmonad status bar
recently the use of dzen has resulted in either the windows covering
dzen or dzen covering the top of the windows. I tried all kinds of
recommended way to fix it, but couldn't. Switching to xmobar for that
side of the status bar works and will hopefully be better supported.